Signup_Module/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
cryptography==3.4.6
cssselect==1.1.0
cssutils==1.0.2
cv==1.0.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1
django-allauth==0.44.0
django-crispy-forms==1.11.0
django-import-export==2.5.0
django-multi-email-field==0.6.2
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
django-premailer==0.2.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
dlib==19.21.1
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
idna==2.10
jdcal==1.4.1
lxml==4.6.2
MarkupPy==1.14
numpy==1.20.1
oauthlib==3.1.0
odfpy==1.4.1
opencv-python==4.5.1.48
openpyxl==3.0.4
phonenumbers==8.12.18
Pillow==8.1.0
premailer==3.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.26.3
virtualenv==20.4.2
xlrd==1.2.0
xlwt==1.3.0